Your Name: Kristen Tonga
Subject Line: Calculator - done, other two - not so much.

What Happened: got the calculator up and running, but did it before the earlier assignments, and it's quite late now. I haven't checked any of the answers for the name generator, and figure I'll work on the coding for it and for the random number guessing game in the morning.