Your Name: Kristen Tonga
Subject Line:

What Happened: (see last commit message). Included is Tues, Wed and Thurs hw. Tuesday is completed (and I worked on creating a calculator that can pass the tests again last night - but it's not quite working yet). On Wed, I have a beautiful smoothie maker, but haven't done the subway assignment yet (weekend hw), and the Thursday hw is 2/3 or 3/4 done saved in the apartment.rb file. The building.rb file is presently empty.